requirements.txt
mysql-python==1.2.5
django==1.8
django-mptt==0.8.5
django-suit==0.2.18

